PortMapperConfigurerTests groovy->java
Issue: gh-4939
